requirements.txt
# Django==2.2.16
# pytz==2018.5

django
djangorestframework
django-rest-passwordreset
psycopg2-binary
pyyaml
ujson
requests
django-filter
pytest-django
pytest
social-auth-app-django
django-silk